Bug fix: close the windows when Xed

+++ RosalilaInputs/Boton.cpp
@@ -48,7 +48,7 @@ std::string convertInt(int number)
         temp+=number%10+48;
         number/=10;
     }
-    for (int i=0;i<temp.length();i++)
+    for (int i=0;i<(int)temp.length();i++)
         returnvalue+=temp[temp.length()-i-1];
     return returnvalue;
 }
